+++ openraft/tests/membership/t30_commit_joint_config.rs
@@ -66,7 +66,7 @@ async fn commit_joint_config_during_0_to_012() -> Result<()> {
         let router = router.clone();
         async move {
             let node = router.get_raft_handle(&0).unwrap();
-            let _x = node.change_membership(btreeset! {0,1,2}, true, false).await;
+            let _x = node.change_membership(btreeset! {0,1,2}, false).await;
         }
         .instrument(tracing::debug_span!("spawn-change-membership"))
     });
@@ -113,7 +113,7 @@ async fn commit_joint_config_during_012_to_234() -> Result<()> {
 
     tracing::info!("--- changing config to 0,1,2");
     let node = router.get_raft_handle(&0)?;
-    node.change_membership(btreeset![0, 1, 2], true, false).await?;
+    node.change_membership(btreeset![0, 1, 2], false).await?;
     log_index += 2;
 
     router.wait_for_log(&btreeset![0, 1, 2], Some(log_index), None, "cluster of 0,1,2").await?;
@@ -125,7 +125,7 @@ async fn commit_joint_config_during_012_to_234() -> Result<()> {
         tokio::spawn(
             async move {
                 let node = router.get_raft_handle(&0)?;
-                node.change_membership(btreeset![2, 3, 4], true, false).await?;
+                node.change_membership(btreeset![2, 3, 4], false).await?;
                 Ok::<(), anyhow::Error>(())
             }
             .instrument(tracing::debug_span!("spawn-change-membership")),
